+++ CHANGELOG.md
@@ -4,8 +4,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Blueprints and templates now have the output item in their name, instead of the tooltip
 ### Fixed
 - Stats not calculating correctly in some cases (e.g. tip upgrades) [#27]
+- Rod and bowstring blueprints/templates being consumed during crafting
 
 ## [1.0.11] - 2019-05-08
 ### Added